feature(server): 实现主键生成

+++ src/test/java/cn/darkjrong/hbase/service/StudentService.java
@@ -2,5 +2,5 @@ package cn.darkjrong.hbase.service;
 
 import cn.darkjrong.hbase.Student;
 
-public interface StudentService extends HbaseService<Student, Integer> {
+public interface StudentService extends HbaseService<Student, Long> {
 }
