+++ src/test/java/cn/darkjrong/hbase/service/impl/StudentServiceImpl.java
@@ -3,7 +3,7 @@ package cn.darkjrong.hbase.service.impl;
 import cn.darkjrong.hbase.Student;
 import cn.darkjrong.hbase.service.StudentService;
 
-public class StudentServiceImpl extends HbaseServiceImpl<Student, Integer> implements StudentService {
+public class StudentServiceImpl extends HbaseServiceImpl<Student, Long> implements StudentService {
